requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
fastapi>=0.110.0
psycopg2-binary>=2.9.0
pytest>=8.0.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0